+++ app/component/map/generic-marker.cjsx
@@ -2,7 +2,7 @@ config = require('../../config')
 React         = require 'react'
 isBrowser     = window?
 Marker        = if isBrowser then require('react-leaflet/lib/Marker').default
-Popup         = if isBrowser then require('./dynamic-popup').default
+Popup         = if isBrowser then require('./Popup').default
 L             = if isBrowser then require 'leaflet'
 provideContext = require 'fluxible-addons-react/provideContext'
 ComponentUsageExample = require('../documentation/ComponentUsageExample').default
