+++ app/component/map/vehicle-marker-container.cjsx
@@ -2,7 +2,7 @@ React         = require 'react'
 Relay         = require 'react-relay'
 queries       = require '../../queries'
 isBrowser     = window?
-Popup  = if isBrowser then require('./dynamic-popup').default
+Popup  = if isBrowser then require('./Popup').default
 RouteMarkerPopup = require './route/route-marker-popup'
 Marker        = if isBrowser then require('react-leaflet/lib/Marker').default else null
 L             = if isBrowser then require 'leaflet' else null
